Adding wait-for-it.sh script in entrypoint to wait for kafka without using time.sleep, Adding default network back into docker-compose, Improving scripts by checking arguments, Adding --deep option to clean.sh

+++ scripts/clean.sh
@@ -5,4 +5,8 @@ docker service rm $(docker service ls -q)
 docker rm $(docker ps -a -q)
 docker container rm $(docker container ls -q -a)
 docker network rm $(docker network ls -q)
-docker image rm -f $(docker image ls -q -a)
+
+if [[ "$1" = "--deep" ]]
+then
+	docker image rm -f $(docker image ls -q -a)
+fi
